refactor: User 도메인으로 패키지 구조 통합 및 auth 관련 클래스 이동
- 기존의 auth 패키지에 분산되어 있던 클래스들을 domain/User 하위로 통합하여
도메인 중심의 구조로 리팩토링하였습니다
- 컨트롤러, DTO, 서비스, 엔티티 등의 클래스들이 User 도메인으로 이동되었으며,
불필요한 중복 클래스들도 함께 정리하였습니다.
