+++ src/main/java/com/project/growfit/domain/User/dto/request/ParentSignUpRequest.java
@@ -1,7 +1,7 @@
 package com.project.growfit.domain.auth.dto.request;
 
 import com.project.growfit.domain.auth.entity.Parent;
-import com.project.growfit.domain.auth.entity.ROLE;
+import com.project.growfit.domain.User.entity.ROLE;
 
 public record ParentSignUpRequest(
         String email,
