--- src/main/java/com/project/growfit/domain/auth/entity/GENDER.java
@@ -1,14 +1,0 @@
-package com.project.growfit.domain.auth.entity;
-
-public enum GENDER {
-    GENDER_MALE,
-    GENDER_FEMALE;
-
-    public static ROLE toEnum(String role) {
-        try {
-            return ROLE.valueOf("GENDER" + role.toUpperCase());
-        } catch (IllegalArgumentException e) {
-            return null;
-        }
-    }
-}
